Design pattern hozzáadása, eddigiek hozzáadása
