+++ Morning_wakeup_app/MainPage.xaml
@@ -8,17 +8,43 @@
     mc:Ignorable="d"
     Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
 
-    <Grid>
-        <Button x:Name="weather_button" Content="Get current weather informations" Margin="72,105,0,0" VerticalAlignment="Top" Click="weather_button_Click"/>
-        <TextBlock x:Name="weather_textblock" HorizontalAlignment="Left" Margin="72,145,0,0" Text="" TextWrapping="Wrap" VerticalAlignment="Top" Height="235" Width="342" RenderTransformOrigin="0.504,0.531"/>
-        <Image x:Name="Weather_img" HorizontalAlignment="Left" Height="100" Margin="501,55,0,0" VerticalAlignment="Top" Width="100"/>
-        <Button x:Name="news_button" Content="Get the news" Margin="72,467,0,0" VerticalAlignment="Top" Width="211" Click="news_button_Click"/>
-        <TextBlock x:Name="news_tb" HorizontalAlignment="Left" Margin="72,505,0,0" Text="" TextWrapping="Wrap" VerticalAlignment="Top" Height="318" Width="639"/>
-        <TextBlock x:Name="time_tb" HorizontalAlignment="Left" Margin="1318,40,0,0" Text="" TextWrapping="Wrap" VerticalAlignment="Top" Height="66" Width="139"/>
-        <Button x:Name="next_news_button" Content="Next news" Margin="465,467,0,0" VerticalAlignment="Top" Click="next_news_button_Click"/>
-        <Button x:Name="prev_news_button" Content="Previous news" Margin="329,467,0,0" VerticalAlignment="Top" Click="prev_news_button_Click"/>
-        <TextBox x:Name="news_search_input_tb" HorizontalAlignment="Left" Margin="72,430,0,0" TextWrapping="Wrap" VerticalAlignment="Top" PlaceholderText="Type here to list the news by" InputScope="Default" TextChanged="news_search_input_tb_TextChanged"/>
-        <TextBox x:Name="weather_search_input_tb" HorizontalAlignment="Left" Margin="72,68,0,0" Text="" TextWrapping="Wrap" VerticalAlignment="Top" PlaceholderText="Type here the city name" TextChanged="weather_search_input_tb_TextChanged"/>
-
+    <Grid Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="*" />
+        </Grid.RowDefinitions>
+        <TextBlock x:Name="time_tb" HorizontalAlignment="Left" Margin="1318,0,0,0" Text="" TextWrapping="Wrap" VerticalAlignment="Top" Height="66" Width="139"/>
+        <Frame x:Name="frame" Background="Azure" Grid.Row="1" Margin="0,0,0,0" />
+        <RelativePanel>
+            <Button Name="testButton" FontFamily="Segoe MDL2 Assets" Content="&#xE700;" FontSize="36" Click="testButton_Click" />
+        </RelativePanel>
+        <SplitView Name="MySplitView"
+                   Grid.Row="1"
+                    DisplayMode="CompactOverlay"
+                   OpenPaneLength="300" 
+                   CompactPaneLength="56" 
+                   HorizontalAlignment="Left">
+            <SplitView.Pane>
+                <ListBox SelectionMode="Single"
+                     Name="IconsListBox" 
+                     SelectionChanged="IconsListBox_SelectionChanged">
+                    <ListBoxItem Name="ShareListBoxItem">
+                        <StackPanel Orientation="Horizontal">
+                            <TextBlock FontFamily="Segoe MDL2 Assets" FontSize="36" Text="&#xE72D;" />
+                            <TextBlock Text="Share" FontSize="24" Margin="20,0,0,0" />
+                        </StackPanel>
+                    </ListBoxItem>
+                    <ListBoxItem Name="FavoritesListBoxItem">
+                        <StackPanel Orientation="Horizontal">
+                            <TextBlock FontFamily="Segoe MDL2 Assets" FontSize="36" Text="&#xE734;" />
+                            <TextBlock Text="Favorites" FontSize="24" Margin="20,0,0,0" />
+                        </StackPanel>
+                    </ListBoxItem>
+                </ListBox>
+            </SplitView.Pane>
+            <SplitView.Content>
+                <TextBlock Name="ResultTextBlock" />
+            </SplitView.Content>
+        </SplitView>
     </Grid>
 </Page>
