+++ Morning_wakeup_app/MainPage.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Morning_wakeup_app.XAML_Pages;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -13,7 +14,6 @@ using Windows.UI.Xaml.Input;
 using Windows.UI.Xaml.Media;
 using Windows.UI.Xaml.Media.Imaging;
 using Windows.UI.Xaml.Navigation;
-using static Morning_wakeup_app.News;
 
 // The Blank Page item template is documented at https://go.microsoft.com/fwlink/?LinkId=402352&clcid=0x409
 
@@ -34,61 +34,18 @@ namespace Morning_wakeup_app
             Second_timer.Start();
         }
 
-        private async void weather_button_Click(object sender, RoutedEventArgs e)
-        {
-            var flag = await Current_weather.GetWeatherInformations();
-            string weather_icon = String.Format("http://openweathermap.org/img/wn/{0}@2x.png", Current_weather.weather_reports.weather[0].icon);
-            Weather_img.Source = new BitmapImage(new Uri(weather_icon, UriKind.Absolute));
-
-            weather_textblock.Text = Current_weather.weather_reports.name + "\n" + Current_weather.weather_reports.main.temp + " C " + "\n" + Current_weather.weather_reports.main.feels_like + " C" + "\n" + Current_weather.weather_reports.main.humidity + " %" +"\n" + Current_weather.weather_reports.main.pressure + " Pa" + "\n" + Current_weather.weather_reports.weather[0].description;
-        }
-
-        private async void news_button_Click(object sender, RoutedEventArgs e)
-        {
-            var flag = await News.GetArticlesMain();
-            News.current_news_index = 0;
-            news_tb.Text = News.news_articles.First().title + "\n" + News.news_articles.First().author + "\n" + News.news_articles.First().description;
-        }
-
         private void Second_timer_Tick(object sender, object e)
         {
             time_tb.Text = DateTime.Now.ToString("h:mm:ss tt");
         }
-
-        private void next_news_button_Click(object sender, RoutedEventArgs e)
+        private void testButton_Click(object sender, RoutedEventArgs e)
         {
-            int length = News.news_articles.Count();
-            if (News.news_articles != null)
-            {
-                if (News.current_news_index+1 != length)
-                    News.current_news_index += 1;
-                news_tb.Text = News.news_articles[News.current_news_index].title + "\n" + News.news_articles[News.current_news_index].author + "\n" + News.news_articles[News.current_news_index].description;
-            }
-                
-
-
+            MySplitView.IsPaneOpen = !MySplitView.IsPaneOpen;
         }
-
-        private void prev_news_button_Click(object sender, RoutedEventArgs e)
+        private void IconsListBox_SelectionChanged(object sender, SelectionChangedEventArgs e)
         {
-            //var flag = await News.GetArticlesMain();
-            if (News.news_articles != null)
-            {
-                if (News.current_news_index > 0)
-                    News.current_news_index -= 1;
-                news_tb.Text = News.news_articles[News.current_news_index].title + "\n" + News.news_articles[News.current_news_index].author + "\n" + News.news_articles[News.current_news_index].description;
-            }
-            
-        }
-
-        private void news_search_input_tb_TextChanged(object sender, TextChangedEventArgs e)
-        {
-            News.search_by = news_search_input_tb.Text;
-        }
-
-        private void weather_search_input_tb_TextChanged(object sender, TextChangedEventArgs e)
-        {
-            Current_weather.search_by = weather_search_input_tb.Text;
+            if (ShareListBoxItem.IsSelected) { frame.Navigate(typeof(Page1)); }
+            else if (FavoritesListBoxItem.IsSelected) { frame.Navigate(typeof(Page2)); }
         }
     }
 }
